sp14up31

+++ controllers/cards.js
@@ -21,8 +21,6 @@ module.exports.getCard = (req, res, next) => {
 
 module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
-  console.log(req.body);
-  console.log(req.user._id);
   Card.create({ name, link, owner: req.user._id })
     // { new: true }
     .then((card) => {
@@ -44,23 +42,17 @@ module.exports.createCard = (req, res, next) => {
 };
 
 module.exports.deleteCard = (req, res, next) => {
-  // console.log(req.params.cardId);
-  // Card.findById(req.params.cardId)
-  //   // { new: true }
-  // .then((data) => {
-  // if (data.owner.valueOf() === req.user._id) {
-  // console.log(data.owner._id.valueOf());
-  // console.log(req.user._id);
-  Card.findByIdAndRemove(req.params.cardId, { new: true })
-    .then((card) => {
-      res.status(200).send({ data: card });
+  Card.findById(req.params.cardId)
+    .then((data) => {
+      if (data.owner.valueOf() === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId, { new: true })
+          .then((card) => {
+            res.status(200).send({ data: card });
+          });
+      } else {
+        next(new ForbiddenError('У вас нет прав для осуществления этого действия.'));
+      }
     })
-    // } else {
-    // console.log(data.owner._id.valueOf());
-    // console.log(req.user._id);
-    //   next(new ForbiddenError('У вас нет прав для осуществления этого действия.'));
-    // }
-    // })
     .catch((err) => {
       if (err.name === 'CastError' || err.name === 'ValidationError') {
         next(new BadRequestError('Переданы некорректные данные.'));
